feat: restructure folder organization
