+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import { ThemeToggle } from './components/ThemeToggle';
-import { LanguageSwitcher } from './components/LanguageSwitcher';
+import { ThemeToggle } from '../components/shared/ThemeToggle';
+import { LanguageSwitcher } from '../components/shared/LanguageSwitcher';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
